requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==6.7
cssselect==0.9.1
cssutils==1.0.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Limiter==0.9.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.5.0
idna==2.7
iniherit==0.3.9
itsdangerous==0.24
Jinja2==2.10.1
limits==1.1
lxml==3.5.0
MarkupSafe==0.23
premailer==2.9.6
psycopg2==2.7.5
pydns==2.3.6
requests==2.20.0
six==1.10.0
SQLAlchemy==1.3.0
text-unidecode==1.3
urllib3==1.24.3
validate-email==1.3
Werkzeug==0.14.1
yoyo-migrations==6.1.0
